chore: rename specbox to spec-box [x] no test file to update
