+++ packages/cfd/src/Components/spec-box.tsx
